project/requirements.txt
appdirs==1.4.4
attrs==23.1.0
autopage==0.5.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cliff==4.3.0
cmd2==2.4.3
cryptography==41.0.4
debtcollector==2.5.0
decorator==5.1.1
dogpile.cache==1.2.2
idna==3.4
importlib-metadata==6.8.0
iso8601==2.1.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
keystoneauth1==5.3.0
msgpack==1.0.7
netaddr==0.9.0
netifaces==0.11.0
openstacksdk==1.5.0
os-service-types==1.7.0
osc-lib==2.8.1
oslo.config==9.2.0
oslo.i18n==6.1.0
oslo.serialization==5.2.0
oslo.utils==6.2.1
packaging==23.2
pbr==5.11.1
prettytable==3.9.0
pycparser==2.21
pyparsing==3.1.1
pyperclip==1.8.2
python-cinderclient==9.4.0
python-dotenv==1.0.0
python-keystoneclient==5.2.0
python-novaclient==18.4.0
python-openstackclient==6.3.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requestsexceptions==1.4.0
rfc3986==2.0.0
simplejson==3.19.2
six==1.16.0
stevedore==5.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.8
wrapt==1.15.0
zipp==3.17.0
